// deploy_backend/src/types/api.types.ts
export interface OrderQuoteRequest {
  orderId: string;
  ownerCode?: string;
  selectedOffers?: SelectedOffer[];
  addServices?: SelectedService[];
  seatSelections?: SeatSelection[];
  distributionChain?: DistributionChain;
}